Move createTreeBuilder to Search::Module and remove subdir
