+++ apptainer/2023-08-09_tensorflow-2.8_onnx-1.15_v1/makefiles/Modules.make
@@ -147,7 +147,6 @@ endif
 
 # ****** Libraries ******
 LIBS_SEARCH = src/Search/libSprintSearch.$(a)
-LIBS_SEARCH += src/Search/TreeBuilder/libSprintTreeBuilder.$(a)
 ifdef MODULE_SEARCH_WFST
 LIBS_SEARCH += src/Search/Wfst/libSprintSearchWfst.$(a)
 LIBS_SEARCH += src/OpenFst/libSprintOpenFst.$(a)
