+++ src/Search/AdvancedTreeSearch/SearchSpaceHelpers.hh
@@ -16,7 +16,7 @@
 #define SEARCH_SPACE_HELPERS
 
 #include <Core/Types.hh>
-#include <Search/TreeBuilder/TreeStructure.hh>
+#include <Search/TreeStructure.hh>
 
 #include "LanguageModelLookahead.hh"
 #include "TraceManager.hh"
