+++ src/Search/LanguageModelLookahead.hh
@@ -25,7 +25,7 @@
 #include <Core/Parameter.hh>
 #include <Core/ReferenceCounting.hh>
 #include <Lm/ScaledLanguageModel.hh>
-#include <TreeBuilder/StateTree.hh>
+#include "StateTree.hh"
 
 namespace Search {
