+++ src/Search/Module.hh
@@ -18,10 +18,19 @@
 #include <Core/Configuration.hh>
 #include <Core/Singleton.hh>
 
+#include "TreeBuilder.hh"
+
 namespace Search {
 
 class SearchAlgorithm;
 class LatticeHandler;
+
+enum class TreeBuilderType {
+    previousBehavior = 0,
+    classicHmm       = 1,
+    minimizedHmm     = 2,
+    ctc              = 3,
+};
 
 enum SearchType {
     WordConditionedTreeSearchType,
@@ -34,8 +43,9 @@ class Module_ {
 public:
     Module_();
 
-    SearchAlgorithm* createRecognizer(SearchType type, const Core::Configuration& config) const;
-    LatticeHandler*  createLatticeHandler(const Core::Configuration& c) const;
+    std::unique_ptr<AbstractTreeBuilder> createTreeBuilder(TreeBuilderType treeBuilderType, Core::Configuration config, const Bliss::Lexicon& lexicon, const Am::AcousticModel& acousticModel, Search::PersistentStateTree& network, bool initialize = true) const;
+    SearchAlgorithm*                     createRecognizer(SearchType type, const Core::Configuration& config) const;
+    LatticeHandler*                      createLatticeHandler(const Core::Configuration& c) const;
 };
 
 typedef Core::SingletonHolder<Module_> Module;
