+++ src/Search/TreeBuilder.hh
@@ -16,10 +16,10 @@
 #define TREEBUILDER_HH
 
 #include <Bliss/Phoneme.hh>
+#include <Search/StateTree.hh>
 #include <Search/AdvancedTreeSearch/Helpers.hh>
 #include <Search/AdvancedTreeSearch/LmCache.hh>
 #include "PersistentStateTree.hh"
-#include "StateTree.hh"
 
 namespace Bliss {
 class Lexicon;
@@ -32,13 +32,6 @@ class AcousticModel;
 namespace Core {
 class Configuration;
 }
-
-enum class TreeBuilderType {
-    previousBehavior = 0,
-    classicHmm       = 1,
-    minimizedHmm     = 2,
-    ctc              = 3,
-};
 
 class AbstractTreeBuilder : public Core::Component {
 public:
@@ -292,18 +285,4 @@ protected:
     void addWordBoundaryStates();
 };
 
-std::unique_ptr<AbstractTreeBuilder> createTreeBuilder(TreeBuilderType treeBuilderType, Core::Configuration config, const Bliss::Lexicon& lexicon, const Am::AcousticModel& acousticModel, Search::PersistentStateTree& network, bool initialize = true) {
-    switch (treeBuilderType) {
-        case TreeBuilderType::classicHmm: {  // Use StateTree.hh
-            return std::unique_ptr<AbstractTreeBuilder>(nullptr);
-        } break;
-        case TreeBuilderType::minimizedHmm: {  // Use TreeStructure.hh
-            return std::unique_ptr<AbstractTreeBuilder>(new MinimizedTreeBuilder(config, lexicon, acousticModel, network, initialize));
-        } break;
-        case TreeBuilderType::ctc: {
-            return std::unique_ptr<AbstractTreeBuilder>(new CtcTreeBuilder(config, lexicon, acousticModel, network, initialize));
-        } break;
-        default: defect();
-    }
-}
 #endif
