+++ src/Search/TreeStructure.hh
@@ -18,8 +18,8 @@
 #include <vector>
 
 #include <Core/MappedArchive.hh>
+#include <Search/StateTree.hh>
 #include <Search/AdvancedTreeSearch/BatchManager.hh>
-#include "StateTree.hh"
 
 #define inline_ __attribute__((always_inline)) inline
