+++ src/Tools/NnTrainer/Makefile
@@ -25,7 +25,6 @@ NN_TRAINER_O = $(OBJDIR)/NnTrainer.o \
 		../../Mm/libSprintMm.$(a) \
 		../../Nn/libSprintNn.$(a) \
 		../../Search/libSprintSearch.$(a) \
-		../../Search/TreeBuilder/libSprintTreeBuilder.$(a) \
 		../../Signal/libSprintSignal.$(a) \
 		../../Speech/libSprintSpeech.$(a)
